requirements.txt
beautifulsoup4==4.12.3
pandas==2.2.2
PyInquirer==1.0.3
pyodbc==5.1.0
python-dotenv==1.0.1
Requests==2.32.3
tqdm==4.66.4
